Get rid of LibGDX.
It's a bad idea to include LibGDX because some projects that depend on
fg-tools don't want to transitively pull in something as heavy LibGDX,
when they aren't even going to use it.

Furthermore, it goes against the goals of the fg-tools project, which is
to be a small, lightweight, general-purpose, game utility library.