Make Result failure message generic.
Result failure messages are limited to String type. String is a great
default, but there are now requirements to be able to use a custom type
in some cases. Result <String> should generally be used as the default
now, instead of just Result (which now results in a 'raw type' warning).

Other Result improvements:

The concept of a "message" has been clarified to mean a reason for
failure. Therefore, when using Result.success(), failureReason will be
null, and an IllegalStateException will be thrown upon attempting to
obtain a failure reason from a successful Result.

Add some methods to help with fluency:

- succeeded() [same as isSuccessful()]
- failed() [same as isFailure()]
- failedBecauseOf (T failureReason) [same as getFailureReason.equals (T failureReason)]

Although failureReason will be null when invoking Result.success(), it
is still required to be non-null when invoking Result.failure (T failureReason).
This helps provide consistency of use; that is, a failure always
requires a valid reason, and a success can never have one. When a result
is a failure, it can be relied upon that there is a non-null reason
associated with it, and vice versa for successful results.

Here is a real-world example of the above changes in use (in a junit test):

  assertTrue (playerModel.requestToSetPlayerLimitTo(newLimit).failedBecauseOf (ChangePlayerLimitDeniedEvent.REASON.CANNOT_DECREASE_BELOW_ZERO));