Make DeniedEvent have a generic reason type.
DeniedEvent previously required a String type to provide the reason for
the denial. String is a great default, but there are now requirements to
be able to use a custom type in some cases. DeniedEvent <String> should
generally be used as the default now, instead of just DeniedEvent (which
now results in a 'raw type' warning).

This goes hand-in-hand with the previous commit regarding generifying
the Result class. DeniedEvent's work well with Result's, as a failure
Result can pass it's reason for failure into a DeniedEvent's reason for
denial. So it makes sense to allow DeniedEvent's to allow custom types
for denial reasons since Result now allows the same for failure reasons.