Add hasName method to Asset interface.
There is already a method 'has (String name)' in the Asset interface,
but it can be confusing in some contexts, e.g.:

  if (player.has ("Bob")) ...

In other cases, this method helps with code fluency:

  String name = "Bob";
  ...
  if (player.has (name)) ...

instead of:

  String name = "Bob";
  ...
  if (player.hasName (name)) ...

which is redunant. The solution to the first case is to add the method
'hasName (String name)':

  if (player.hasName ("Bob")) ...

which is NOT redundant and makes better sense. Both of these methods are
useful in certain contexts, as demonstrated above.