Remove uploadArchives gradle task.
uploadArchives has been configured to publish to a github-based maven
repository. The uploading works; unfortunately, the downloading no
longer does. There is no point to uploading to a maven repository that
can't be accessed by projects that depend on fg-tools. So, remove it.

This means that projects depending on fg-tools will need to clone the
project from github, and build it manually in order for it to be
accessible in the local maven cache.