Remove logback.
It was added due to a misunderstanding about it's use. Since fg-tools is
a library that does not do anything itself, clients / users of the
library are supposed to specify the logging implementation. Slf4j is the
flexible logging framework / api, that allows any logging framework that
implements the api to be used.

The whole point of slf4j is for the client / user to be able to choose
which logging implementation they want to use. By specifying logback as
the implementation in this library, users are either forced into using
logback, or must exclude the transitive dependency.