Make Randomness work on Android.
On some platforms, such as Android, the "SUN" security Provider is not
available. Because Randomness explicitly specifies the "SUN" security
provider as a "best practices" technique, SecureRandom creation will
fail with a NoSuchProviderException.

A more robust solution in such cases is to fall back to the "preferred"
provider for the platform, which will also be the most
cryptographically-strong, which is the best fall-back strategy that can
be managed.