Make DefaultExternalAddressResolver more robust.
In DefaultExternalAddressResolver#resolveIp, check that reading a line
from the provided URL does not return null, and check that it is a valid
ip address string using InetAddresses#isInetAddress. If not, in both
cases throw an IOException explaining that the ip is invalid, which will
be caught in the catch block, and logged as a warning, and an empty ip
string will be returned.