Add "membership" check to RecyclableobjectPool.release.
There was previously no check in the release() method in RecyclableObjectPool for whether or not the released object actually exists in the pool's checkout reference cache.

That means it would have been possible to manually add objects to the pool using "release" as a backdoor.

Think of all the possible hacks...