Fix bug in Resources#getJarChildDirectoryNames.
While the last commit claimed:

"Allow a forward slash to be prepended to the beginning of the specified
parent directory path, instead of prepending one automatically, as some
callers need to pass in a path that is already prepended with a forward
slash."

The forward slash was still being prepended automatically. Remove it.