Add net::isValidAddress method.
Checks whether the specified nullable address string is a valid domain
name, ipv6, or ipv4 string.

No network lookups are performed in this method; it is strictly offline
syntax checking, so even if this method returns true, there is no
guarantee that the specified address actually exists and/or is
accessible.