Improve gradle subproject naming scheme.
Currently, gradle uses the folder name as the name of a subproject's
archive, currently resulting in:

parent project: 'fg-tools'
  subproject: 'common'
  subproject: 'net'

This problem becomes apparent when deploying or using the library's
artifacts, as the jar files will simply be named 'common' & 'net'
(ignoring versioning), creating very potential naming confusion,
especially if deploying to, for example, maven central. There is nothing
in a subproject's name as it currently is to indicate that it is part of
the fg-tools project.

Prefix the root project name, 'fg-tools', to each subproject archive
using gradle's archiveBaseName property.

The new subproject archive names are fg-tools-common & fg-tools-net.
While more verbose, it is a compromise that eliminates confusion and
potential naming conflicts with other dependencies when this library is
used in other projects.

P.S. This is not a package namespace issue, as the subprojects are
already in packages: com.forerunnergames.tools.common &
com.forerunnergames.tools.net.