Add MutatorResult.
MutatorResult allows for deferred execution of successful mutation operations. For example, let's say some verification code  for an object mutator method runs on some given input and determines that it is valid. The function can return a successful MutatorResult with a local/anonymous MutatorCallback type that will perform the actual data update when commitChanges() is called.

This is useful for cases when multiple operations need to be validated and none are allowed to fail. The actual updates will only occur if and only if the user calls commitChanges() on each MutatorResult (or uses MutatorResult.commitAllSuccessful).