Fix delayed cyclic barrier test.
DelayedCyclicBarrier FIFO test was using a delay of 1 ms which is too short to reasonably guarantee any level of reliability in subsequent thread execution order.

In the span of a 1 ms delay, another thread or process outside of the test thread pool might be executed. Upon return, there is no guarantee which thread in the pool will be selected for execution.

DelayedCyclicBarrier is meant to provide only a very basic measure of thread execution ordering after a join. It cannot be expected to behave reliably within such a small finitude of time.

Repeated testing indicates that a delay of 5 ms seems to be far more reliable. Sporadic test failures may still be possible in the event of JVM or system runtime anomalies; i.e. the test process gets pre-empted and queued for > 5 ms. But this is very unlikely.

PERIL-753 #Done #time 10m